make immediate webpackable by not referencing global as global variable
fixes #39

for usage in webpack environments when there is no chance for polyfills, the actual implementation causes crashes, as global is not defined.
the browserified version in the dist directory is not usable for webpack and especially for treeshaking. To fix this, global is not referenced anymore as global variable, it is passed to every function requiring it and the "global polyfill" is prepended to the index.js